Use usize fo the dimension variable

+++ src/bad_guy.rs
@@ -6,14 +6,14 @@ pub struct BadGuy {
 } 
 
 impl BadGuy {
-    pub fn new(dimension: i8) -> BadGuy {
+    pub fn new(dimension: usize) -> BadGuy {
         let mut rng = rand::thread_rng();
         let num = Uniform::from(0..dimension);
         let x = num.sample(&mut rng);
         let y = num.sample(&mut rng);
         println!("bad guy at: {}, {}", x, y);
 
-        let position = (x as usize, y as usize);
+        let position = (x, y);
 
         BadGuy{position, is_active: false}
     }
